Implement data persistency (#17)
# Implement data persistency

Until now Hermes could work only as an in-memory storage and when it was restarted, all it's data was gone. By this implementation, a local append file method is created where the set and remove type actions are logged. When Hermes is starting, this file is read and instructions are executed.

Hermes remains in-memory storage, this implementation writes log asyncornousily, so it does not reduce the performance. Logging can be turned off from config.toml file.

## Implementaiton plan

- [x] Implement append file
- [x] Write test
- [x] Review documenation
- [x] Review docker

Reviewed-on: https://git.thinkaboutit.tech/PublicProjects/olympus-hermes/pulls/17
Co-authored-by: Attila Molnar <molnar.attila28@gmail.com>
Co-committed-by: Attila Molnar <molnar.attila28@gmail.com>